image_processing_cf/requirements.txt
pydantic
functions-framework
requests
google-cloud-aiplatform
google-cloud-storage